moved Connection related components into "timetable/"
