+++ src/app/components/timetable/connection/WingTrain.tsx
@@ -1,6 +1,5 @@
-import React from "react";
 import { Journey } from "@/app/lib/objects";
-import ScheduledComponent from "@/app/components/scheduled";
+import ConnectionComponent from "@/app/components/timetable/connection/Connection";
 
 interface Props {
     journey: Journey;
@@ -12,7 +11,7 @@ const WingTrain = ({ journey, isDeparture }: Props) => {
         <div className="flex flex-col">
             {journey.connections.map((connection, index) => (
                 <div key={`connection${index}`}>
-                    <ScheduledComponent
+                    <ConnectionComponent
                         connection={connection}
                         isDeparture={isDeparture}
                         renderBorder={index === 0} // first renders border
